LayoutTests/platform/win/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/type-change-state-expected.txt


PASS change state from hidden to text
PASS change state from hidden to search
PASS change state from hidden to tel
FAIL change state from hidden to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from hidden to email
PASS change state from hidden to password
FAIL change state from hidden to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from hidden to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from hidden to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from hidden to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from hidden to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from hidden to number
PASS change state from hidden to range
FAIL change state from hidden to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from hidden to checkbox
PASS change state from hidden to radio
PASS change state from hidden to file
PASS change state from hidden to submit
PASS change state from hidden to image
PASS change state from hidden to reset
PASS change state from hidden to button
FAIL change state from text to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from text to search
PASS change state from text to tel
FAIL change state from text to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from text to email
PASS change state from text to password
FAIL change state from text to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from text to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from text to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from text to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from text to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from text to number
PASS change state from text to range
FAIL change state from text to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from text to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from text to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from search to text
PASS change state from search to tel
FAIL change state from search to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from search to email
PASS change state from search to password
FAIL change state from search to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from search to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from search to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from search to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from search to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from search to number
PASS change state from search to range
FAIL change state from search to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from search to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from search to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from tel to text
PASS change state from tel to search
FAIL change state from tel to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from tel to email
PASS change state from tel to password
FAIL change state from tel to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from tel to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from tel to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from tel to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from tel to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from tel to number
PASS change state from tel to range
FAIL change state from tel to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from tel to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from tel to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from url to text
PASS change state from url to search
PASS change state from url to tel
PASS change state from url to email
PASS change state from url to password
FAIL change state from url to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from url to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from url to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from url to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from url to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from url to number
PASS change state from url to range
FAIL change state from url to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from url to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from url to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from email to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to text assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from email to search assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from email to tel assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
PASS change state from email to url
FAIL change state from email to password assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "foobar"
FAIL change state from email to datetime assert_equals: input.value should be  after change of state expected "" but got "foobar"
FAIL change state from email to date assert_equals: input.value should be  after change of state expected "" but got "foobar"
FAIL change state from email to month assert_equals: input.value should be  after change of state expected "" but got "foobar"
FAIL change state from email to week assert_equals: input.value should be  after change of state expected "" but got "foobar"
FAIL change state from email to time assert_equals: input.value should be  after change of state expected "" but got "foobar"
PASS change state from email to number
PASS change state from email to range
FAIL change state from email to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "foobar"
FAIL change state from email to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from email to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "foobar"
FAIL change state from password to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from password to text
PASS change state from password to search
PASS change state from password to tel
FAIL change state from password to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from password to email
FAIL change state from password to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from password to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from password to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from password to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from password to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from password to number
PASS change state from password to range
FAIL change state from password to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from password to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from password to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from datetime to text
PASS change state from datetime to search
PASS change state from datetime to tel
FAIL change state from datetime to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from datetime to email
PASS change state from datetime to password
FAIL change state from datetime to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from datetime to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from datetime to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from datetime to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from datetime to number
PASS change state from datetime to range
FAIL change state from datetime to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from datetime to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from datetime to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from date to text
PASS change state from date to search
PASS change state from date to tel
FAIL change state from date to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from date to email
PASS change state from date to password
FAIL change state from date to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from date to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from date to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from date to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from date to number
PASS change state from date to range
FAIL change state from date to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from date to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from date to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from month to text
PASS change state from month to search
PASS change state from month to tel
FAIL change state from month to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from month to email
PASS change state from month to password
FAIL change state from month to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from month to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from month to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from month to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from month to number
PASS change state from month to range
FAIL change state from month to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from month to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from month to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from week to text
PASS change state from week to search
PASS change state from week to tel
FAIL change state from week to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from week to email
PASS change state from week to password
FAIL change state from week to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from week to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from week to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from week to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from week to number
PASS change state from week to range
FAIL change state from week to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from week to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from week to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from time to text
PASS change state from time to search
PASS change state from time to tel
FAIL change state from time to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from time to email
PASS change state from time to password
FAIL change state from time to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from time to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from time to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from time to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from time to number
PASS change state from time to range
FAIL change state from time to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
FAIL change state from time to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from time to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from number to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to text assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to search assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to tel assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
FAIL change state from number to url assert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from number to email assert_equals: input.value should be foobar after change of state expected "foobar" but got ""
FAIL change state from number to password assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got ""
PASS change state from number to datetime
PASS change state from number to date
PASS change state from number to month
PASS change state from number to week
PASS change state from number to time
PASS change state from number to range
FAIL change state from number to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got ""
FAIL change state from number to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from number to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got ""
FAIL change state from range to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to text assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to search assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to tel assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "50"
FAIL change state from range to email assert_equals: input.value should be foobar after change of state expected "foobar" but got "50"
FAIL change state from range to password assert_equals: input.value should be   foobar   after change of state expected "  foobar  " but got "50"
FAIL change state from range to datetime assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to date assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to month assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to week assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to time assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to number assert_equals: input.value should be  after change of state expected "" but got "50"
FAIL change state from range to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "50"
FAIL change state from range to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from range to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "50"
FAIL change state from color to hidden assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from color to text
PASS change state from color to search
PASS change state from color to tel
FAIL change state from color to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from color to email
PASS change state from color to password
FAIL change state from color to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from color to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from color to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from color to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from color to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from color to number
PASS change state from color to range
FAIL change state from color to checkbox assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to radio assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to file assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to submit assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to image assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to reset assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
FAIL change state from color to button assert_equals: input.value should be '  foo\rbar  ' after change of state expected "  foo\rbar  " but got "  foobar  "
PASS change state from checkbox to hidden
PASS change state from checkbox to text
PASS change state from checkbox to search
PASS change state from checkbox to tel
FAIL change state from checkbox to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from checkbox to email
PASS change state from checkbox to password
FAIL change state from checkbox to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from checkbox to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from checkbox to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from checkbox to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from checkbox to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from checkbox to number
PASS change state from checkbox to range
FAIL change state from checkbox to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from checkbox to radio
PASS change state from checkbox to file
PASS change state from checkbox to submit
PASS change state from checkbox to image
PASS change state from checkbox to reset
PASS change state from checkbox to button
PASS change state from radio to hidden
PASS change state from radio to text
PASS change state from radio to search
PASS change state from radio to tel
FAIL change state from radio to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from radio to email
PASS change state from radio to password
FAIL change state from radio to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from radio to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from radio to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from radio to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from radio to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from radio to number
PASS change state from radio to range
FAIL change state from radio to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from radio to checkbox
PASS change state from radio to file
PASS change state from radio to submit
PASS change state from radio to image
PASS change state from radio to reset
PASS change state from radio to button
PASS change state from file to hidden
PASS change state from file to text
PASS change state from file to search
PASS change state from file to tel
PASS change state from file to url
PASS change state from file to email
PASS change state from file to password
PASS change state from file to datetime
PASS change state from file to date
PASS change state from file to month
PASS change state from file to week
PASS change state from file to time
PASS change state from file to number
PASS change state from file to range
PASS change state from file to color
PASS change state from file to checkbox
PASS change state from file to radio
PASS change state from file to submit
PASS change state from file to image
PASS change state from file to reset
PASS change state from file to button
PASS change state from submit to hidden
PASS change state from submit to text
PASS change state from submit to search
PASS change state from submit to tel
FAIL change state from submit to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from submit to email
PASS change state from submit to password
FAIL change state from submit to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from submit to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from submit to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from submit to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from submit to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from submit to number
PASS change state from submit to range
FAIL change state from submit to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from submit to checkbox
PASS change state from submit to radio
PASS change state from submit to file
PASS change state from submit to image
PASS change state from submit to reset
PASS change state from submit to button
PASS change state from image to hidden
PASS change state from image to text
PASS change state from image to search
PASS change state from image to tel
FAIL change state from image to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from image to email
PASS change state from image to password
FAIL change state from image to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from image to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from image to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from image to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from image to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from image to number
PASS change state from image to range
FAIL change state from image to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from image to checkbox
PASS change state from image to radio
PASS change state from image to file
PASS change state from image to submit
PASS change state from image to reset
PASS change state from image to button
PASS change state from reset to hidden
PASS change state from reset to text
PASS change state from reset to search
PASS change state from reset to tel
FAIL change state from reset to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from reset to email
PASS change state from reset to password
FAIL change state from reset to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from reset to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from reset to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from reset to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from reset to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from reset to number
PASS change state from reset to range
FAIL change state from reset to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from reset to checkbox
PASS change state from reset to radio
PASS change state from reset to file
PASS change state from reset to submit
PASS change state from reset to image
PASS change state from reset to button
PASS change state from button to hidden
PASS change state from button to text
PASS change state from button to search
PASS change state from button to tel
FAIL change state from button to url assert_equals: input.value should be foobar after change of state expected "foobar" but got "  foobar  "
PASS change state from button to email
PASS change state from button to password
FAIL change state from button to datetime assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from button to date assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from button to month assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from button to week assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
FAIL change state from button to time assert_equals: input.value should be  after change of state expected "" but got "  foobar  "
PASS change state from button to number
PASS change state from button to range
FAIL change state from button to color assert_equals: input.value should be #000000 after change of state expected "#000000" but got "  foobar  "
PASS change state from button to checkbox
PASS change state from button to radio
PASS change state from button to file
PASS change state from button to submit
PASS change state from button to image
PASS change state from button to reset
